Avoid displaying messages without request in requests feed
This section if for displaying latest messages to specific requests. It
displays the request title as a link. We could display some truncated
text of the message with no request, but then we'd need to account for
messages with a blank text, like an image only message.

I think the section heading is clear that these are the most recent
replies to a specific request