Remove ability to toggle otp enabled state
Closes #2102

This was causing issues because of a lack of undestanding as to why we
added this feature in the first place. We intended on giving admin the
ability to reset someone's otp in the case that there 2fa device was
lost or stolen. However, in practice, admin sometimes toggle this when
creating a user, or helping to debug log in issues. In the event that a
2fa device is lost or stolen, the admin team must open a ticket with the
dev team to reset the otp_enable/otp_secret_key