Enforce non-admin users to have at least 1 org
Fixes #2073

We do not have anything for a non-admin user to do on our site if they do not belong to a single organization.
This prevents creation of a non-admin user without an organization. It
throws an error to make clear to the admin why the user creation failed.
It also adds a help text explaining that a non-admin user must have an
organization assigned in the create user form.