Add tests for quote reply previous request (#2124)
With https://github.com/tactilenews/100eyes/pull/2123, the changes
opened the possibility for properly assigning an incoming message to the
correct request. That was the idea behind
https://github.com/tactilenews/100eyes/pull/2089/commits/5da7eb560a391990a25b51b1bd6b85b010862dec
was to introduce this feature, but it didn't seem to work because we
didn't seem to receive the id in the incoming webhook's payload. I found
that strange since we do for a quote reply/response with a quick reply
button to a template.

I reached out to 360dialog support and the AI said that in fact we do
not receive the id of the message they are quote replying.

However, while working on the reactions, I saw we do receive it for
reactions and double checked it for quote replies and we do seem to. Not
sure if we do in every case, or if something has changed, but I tested
this manually as well and it works as expected.