Fix incorrect request recipients stats
In the RequestMetrics and RequestScheduled the recipients stat was
  calculating based on the current number of active contributors that
the request would go out if created now; however, we are not interested
in that number here. What we are interested in is how many people it
actually went out to. Before creating a new request, it is interesting
to know how many it will go out to, but not after it is already sent.

This PR also unifies the API for this stat since it is used in three
separate places and this is the second time we have had a bug reported
that the values were not in sync.